Refactor Beans: UserBean for general operations (parse ID to name, etc.), AdminBean for admin-related-stuff (list users, create new user) --> only Roles.ADMIN allowed!
